extractor: fix alignment of nested tables
This commit fixes issues caused when camelot parses a table with nested
tables that aren't perfectly aligned. This creates a table like this:

x x - - - -
- - x - x -
x x - - - -
- - - x - x

The code transforms the above table to look like this instead:

x x - -
- - x x
x x - -
- - x x

This makes sure that no data is lost when parsing.

An example of a table in the spec that previously caused these issues is
Figure 259 from Base Specification 2.0b

Signed-off-by: Karl Bonde Torp <k.torp@samsung.com>